+++ bindings/vue/examples/components/index.js
@@ -8,4 +8,4 @@ export { default as Fab } from './Fab.vue';
 export { default as Dialogs } from './Dialogs.vue';
 export { default as Carousel } from './Carousel.vue';
 export { default as Forms } from './Forms.vue';
-export { default as LazyRepeat } from './LazyRepeat.vue';
+export { default as InfiniteScroll } from './InfiniteScroll.vue';