+++ bindings/vue/src/mixins/index.js
@@ -1,3 +1,3 @@
 export { deriveDBB, deriveHandler, deriveEvents } from './derive';
-export { hidable, hasOptions, modifier, selfProvider, dialogCancel, portal } from './common';
+export { hidable, hasOptions, selfProvider, dialogCancel, portal } from './common';
 export { modelInput, modelCheckbox, modelRadio } from './model';